+++ py/torch_tensorrt/dynamo/__init__.py
@@ -1,5 +1,5 @@
 from ._TorchTensorRTModule import TorchTensorRTModule
 from torch_tensorrt.dynamo import fx_ts_compat
-from .backend import compile
 from ._settings import *
-from torch_tensorrt.dynamo import export
+from torch_tensorrt.dynamo import backend
+from torch_tensorrt.dynamo.export import compile
