+++ py/torch_tensorrt/dynamo/common/__init__.py
@@ -1,3 +1,2 @@
-from ._settings import CompilationSettings
 from .input_tensor_spec import InputTensorSpec
 from .fx2trt import TRTInterpreter, TRTInterpreterResult
